[WIP][refactor] Se mueven funciones de servicios a selectores
