+++ events/selectors.py
@@ -1,10 +1,14 @@
-from datetime import time
+import googlemaps
 
+from collections import OrderedDict
+from datetime import datetime, time
+from operator import itemgetter
+
+from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.db.models import Q, QuerySet
 
 from . import models
-from .models import Message
 
 User = get_user_model()
 
@@ -44,6 +48,64 @@ class EventSelector:
             event_enrollments__created_by=attendee, event_enrollments__status=status).order_by('title')
         return event_enrolled_accepted
 
+    def nearby_events_distance(self, user, distance, latitude, longitude, **kwargs):
+        now = datetime.now()
+        filters = {key: val for key, val in kwargs.items() if val}
+        events = self.base_search_events(user).filter(**filters)[:20]
+        result = []
+
+        if events:
+            events_distances_oredered = self.common_method_distance_order(
+                events, latitude, longitude)
+
+            for event, eventdistance in events_distances_oredered.items():
+                if eventdistance <= int(distance):
+                    result.append(event.id)
+                else:
+                    break
+        return models.Event.objects.filter(pk__in=result)
+
+    def events_filter_search(self, user: User, **kwargs) -> QuerySet:
+        now = datetime.now()
+        events = self.base_search_events(user)
+        filters = {key: val for key, val in kwargs.items() if val}
+        return events.filter(**filters)
+
+    def base_search_events(self, user: User) -> QuerySet:
+        now = datetime.now()
+        if user.is_authenticated:
+            events = models.Event.objects.filter(
+                ~Q(event_enrollments__created_by=user) & (Q(start_day__gte=now.date(), start_time__gte=now.time()) | Q(start_day__gte=now.date())))
+        else:
+            events = models.Event.objects.filter(Q(start_day__gte=now.date(
+            ), start_time__gte=now.time()) | Q(start_day__gte=now.date()))
+        return events
+
+    def common_method_distance_order(self, events, latitude, longitude):
+        gmaps = googlemaps.Client(key=settings.GOOGLE_API_KEY)
+
+        latitude_user = latitude
+        longitude_user = longitude
+
+        origins = [{"lat": latitude_user, "lng": longitude_user}]
+        destinations = ""
+
+        for event in events:
+            destinations = destinations + str(
+                event.location_number) + " " + event.location_street + ", " + event.location_city + "|"
+
+        distancematrix = gmaps.distance_matrix(origins, destinations)
+        events_distances = {}
+
+        for element, event in zip(distancematrix['rows'][0]['elements'], events):
+            if element['status'] == 'OK':
+                events_distances[event] = element['distance']['value']
+
+        events_distances_oredered = OrderedDict(
+            sorted(events_distances.items(), key=itemgetter(1), reverse=False))
+
+        return events_distances_oredered
+
 
 class RatingSelector:
     def on_user(self, reviewed: User, on='HOST') -> QuerySet:
@@ -81,5 +143,5 @@ class TransactionSelector:
 
 
 class MessageSelector:
-    def last_message(self) -> Message:
+    def last_message(self) -> models.Message:
         return models.Message.objects.latest()
